% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arb_bin.R
\name{arb_bin}
\alias{arb_bin}
\title{Monotonic binning based on the decision tree}
\usage{
arb_bin(data, y, x)
}
\arguments{
\item{data}{A input dataframe}

\item{y}{The name of Y with 0/1 binary values}

\item{x}{The name of X with numeric values}
}
\value{
A list of binning outcomes, including a list of cut points and a summary dataframe
}
\description{
The function \code{arb_bin} implements the monotonic binning based on the decision tree
by calling the Rborist() function in the Rborist package.
}
\examples{
data(hmeq)
arb_bin(hmeq, BAD, DEROG)
}
