% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bad_bin.R
\name{bad_bin}
\alias{bad_bin}
\title{Monotonic binning by quantile with cases Y = 1}
\usage{
bad_bin(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector with 0/1 binary values}
}
\value{
A list of binning outcomes, including a numeric vector with cut
        points and a dataframe with binning summary
}
\description{
The function \code{bad_bin} implements the quantile-based monotonic binning 
by the iterative discretization based on cases with Y = 1.
}
\examples{
data(hmeq)
bad_bin(hmeq$DEROG, hmeq$BAD)
}
