% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mob.R
\name{bitset_choose}
\alias{bitset_choose}
\title{Retain only a given number of bits.}
\usage{
bitset_choose(ptr, rngs, k, ..., system = getOption("mob.system"))
}
\description{
Retain only a given number of bits.
}
