% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mob.R
\name{bitset_equal}
\alias{bitset_equal}
\title{Returns true if the two bitsets are equal.}
\usage{
bitset_equal(left, right, ..., system = getOption("mob.system"))
}
\description{
Returns true if the two bitsets are equal.
}
