% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mob.R
\name{bitset_insert}
\alias{bitset_insert}
\title{Insert an R integer vector into a bitset.}
\usage{
bitset_insert(ptr, values, ..., system = getOption("mob.system"))
}
\description{
Insert an R integer vector into a bitset.
}
