% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mob.R
\name{bitset_invert}
\alias{bitset_invert}
\title{Perform an in-place negation of a bitset.}
\usage{
bitset_invert(ptr, ..., system = getOption("mob.system"))
}
\description{
Perform an in-place negation of a bitset.
}
