% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mob.R
\name{bitset_to_vector}
\alias{bitset_to_vector}
\title{Convert the bitset into an R integer vector.}
\usage{
bitset_to_vector(ptr, ..., system = getOption("mob.system"))
}
\description{
Convert the bitset into an R integer vector.
}
