% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mob.R
\name{create_wrapper}
\alias{create_wrapper}
\title{Create a wrapper function that dispatches to either the device or host implementation.}
\usage{
create_wrapper(name)
}
\description{
This produces a wrapper function that will call either \verb{\{name\}_host} or
\verb{\{name\}_device}. All arguments are forwarded to the underlying function.
}
\details{
The wrapper function is generated using metaprogramming, allowing it to have
an explicit list of formal parameter names. If we just used \code{...} to forward
argument the argument names wouldn't show up in the documentation.
}
