% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mob.R
\name{double_vector_lround}
\alias{double_vector_lround}
\title{Round values of the vector to the nearest integer.}
\usage{
double_vector_lround(values)
}
\description{
Round values of the vector to the nearest integer.
}
