% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_bin.R
\name{gbm_bin}
\alias{gbm_bin}
\title{Monotonic binning based on the generalized boosted regression model (GBM)}
\usage{
gbm_bin(data, y, x)
}
\arguments{
\item{data}{A input dataframe}

\item{y}{The name of Y with 0/1 binary values}

\item{x}{The name of X with numeric values}
}
\value{
A list of binning outcomes, including a list of cut points and a summary dataframe
}
\description{
The function \code{gbm_bin} implements the monotonic binning based on the generalized boosted 
regression model (GBM) by calling the gbm library.
}
\examples{
data(hmeq)
gbm_bin(hmeq, BAD, DEROG)
}
