% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{integer_vector_match_eq}
\alias{integer_vector_match_eq}
\title{Find indices of elements equal to the given value.}
\usage{
integer_vector_match_eq(vector, value)
}
\description{
Find indices of elements equal to the given value.
}
