% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_bin.R
\name{iso_bin}
\alias{iso_bin}
\title{Monotonic binning based on isotonic regression}
\usage{
iso_bin(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector with 0/1 binary values}
}
\value{
A list of binning outcomes, including a numeric vector with cut
        points and a dataframe with binning summary
}
\description{
The function \code{iso_bin} implements the monotonic binning based on 
the isotonic regression.
}
\examples{
data(hmeq)
iso_bin(hmeq$DEROG, hmeq$BAD)
}
