% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmn_bin.R
\name{kmn_bin}
\alias{kmn_bin}
\title{Monotonic binning based on k-means clustering}
\usage{
kmn_bin(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector with 0/1 binary values}
}
\value{
A list of binning outcomes, including a numeric vector with cut
        points and a dataframe with binning summary
}
\description{
The function \code{kmn_bin} implements the monotonic binning based on 
the k-means clustering
}
\examples{
data(hmeq)
kmn_bin(hmeq$DEROG, hmeq$BAD)
}
