% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_bin.R
\name{qtl_bin}
\alias{qtl_bin}
\title{Monotonic binning by quantile}
\usage{
qtl_bin(data, y, x)
}
\arguments{
\item{data}{A input dataframe}

\item{y}{The name of Y with 0/1 binary values}

\item{x}{The name of X with numeric values}
}
\value{
A list of binning outcomes, including a list of cut points and a summary dataframe
}
\description{
The function \code{qtl_bin} implements the quantile-based monotonic binning by the iterative discretization
}
\examples{
data(hmeq)
qtl_bin(hmeq, BAD, DEROG)
}
