% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mob.R
\name{random_create}
\alias{random_create}
\title{Create a parallel random number generator.}
\usage{
random_create(size, seed = NULL, ..., system = getOption("mob.system"))
}
\description{
Create a parallel random number generator.
}
