% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng_bin.R
\name{rng_bin}
\alias{rng_bin}
\title{Monotonic binning by quantile based on value range}
\usage{
rng_bin(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector with 0/1 binary values}
}
\value{
A list of binning outcomes, including a numeric vector with cut
        points and a dataframe with binning summary
}
\description{
The function \code{rng_bin} implements the quantile-based monotonic binning 
by the iterative discretization based on the equal-width range of values.
}
\examples{
data(hmeq)
rng_bin(hmeq$DEROG, hmeq$BAD)
}
