% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arb_bin.R
\name{arb_bin}
\alias{arb_bin}
\title{Monotonic binning based on decision tree model}
\usage{
arb_bin(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector with 0/1 binary values}
}
\value{
A list of binning outcomes, including a numeric vector with cut
        points and a dataframe with binning summary
}
\description{
The function \code{arb_bin} implements the monotonic binning based on 
the decision tree.
}
\examples{
data(hmeq)
arb_bin(hmeq$DEROG, hmeq$BAD)
}
