% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bad_bin.R
\name{bad_bin}
\alias{bad_bin}
\title{Monotonic binning based on bads only}
\usage{
bad_bin(data, y, x)
}
\arguments{
\item{data}{A input dataframe}

\item{y}{The name of Y with 0/1 binary values}

\item{x}{The name of X with numeric values}
}
\value{
A list of binning outcomes, including a list of cut points and a summary dataframe
}
\description{
The function \code{bad_bin} implements the monotonic binning only based on bads, e.g. Y = 1,
by the iterative discretization.
}
\examples{
data(hmeq)
bad_bin(hmeq, BAD, DEROG)
}
