% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mob.R
\name{bitset_sample}
\alias{bitset_sample}
\title{Retain only a random portion of bits, using on independent bernoulli trials.}
\usage{
bitset_sample(ptr, rngs, p, ..., system = getOption("mob.system"))
}
\description{
Retain only a random portion of bits, using on independent bernoulli trials.
}
