% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_bin.R
\name{gbm_bin}
\alias{gbm_bin}
\title{Monotonic binning based on generalized boosted model}
\usage{
gbm_bin(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector with 0/1 binary values}
}
\value{
A list of binning outcomes, including a numeric vector with cut
        points and a dataframe with binning summary
}
\description{
The function \code{gbm_bin} implements the monotonic binning based on 
the generalized boosted model (GBM).
}
\examples{
data(hmeq)
gbm_bin(hmeq$DEROG, hmeq$BAD)
}
