% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mob.R
\name{integer_vector_match_gt}
\alias{integer_vector_match_gt}
\title{Find indices of elements greaher than the given value.}
\usage{
integer_vector_match_gt(v, value)
}
\description{
Find indices of elements greaher than the given value.
}
