% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_bin.R
\name{iso_bin}
\alias{iso_bin}
\title{Monotonic binning based on isotonic regression}
\usage{
iso_bin(data, y, x)
}
\arguments{
\item{data}{A input dataframe}

\item{y}{The name of Y with 0/1 binary values}

\item{x}{The name of X with numeric values}
}
\value{
A list of binning outcomes, including a list of cut points and a summary dataframe
}
\description{
The function \code{iso_bin} implements the monotonic binning based on the isotonic regression
by calling the isoreg() function in the stat library.
}
\examples{
data(hmeq)
iso_bin(hmeq, BAD, DEROG)
}
